% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{addPredictorCovariate}
\alias{addPredictorCovariate}
\title{Add a predictor for a covariate}
\usage{
addPredictorCovariate(.Object, name, lambdas, len, nameCovariate)
}
\arguments{
\item{.Object}{is the reference of the object in the method}
}
\value{
Return the updated object.
}
\description{
Add a predictor for a covariate
}
