% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{addPredictorSeasonality,Model-method}
\alias{addPredictorSeasonality,Model-method}
\title{gapCV = 24*7, nel nostro caso una settimana
nFold = 5
confidence = 0.95, per le prove non conviene settare l'intervallo di confidenza}
\usage{
\S4method{addPredictorSeasonality}{Model}(
  .Object,
  name,
  nameSeasonality,
  segments,
  sKnots,
  lambdas,
  len
)
}
\description{
gapCV = 24*7, nel nostro caso una settimana
nFold = 5
confidence = 0.95, per le prove non conviene settare l'intervallo di confidenza
}
