% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{addPredictorSeasonality}
\alias{addPredictorSeasonality}
\title{Add a predicttor for a seasonality}
\usage{
addPredictorSeasonality(
  .Object,
  name,
  nameSeasonality,
  segments,
  sKnots,
  lambdas,
  len
)
}
\arguments{
\item{.Object}{is the reference of the object in the method}

\item{split}{What to split on.}
}
\value{
Return the updated object.
}
\description{
Add a predicttor for a seasonality
}
