% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{addPredictorTrend}
\alias{addPredictorTrend}
\title{Add the predictor trend in the wrapper}
\usage{
addPredictorTrend(.Object, name, lambdas, len)
}
\arguments{
\item{.Object}{is the reference of the object in the method}
}
\value{
Return the updated object.
}
\description{
Add the predictor trend in the wrapper
}
