% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{betaPlots}
\alias{betaPlots}
\title{Print and store the plot of beta parameters}
\usage{
betaPlots(.Object)
}
\arguments{
\item{.Object}{is the reference of the object in the method}
}
\value{
Nothing.
}
\description{
Print and store the plot of beta parameters
}
