% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{exportPython}
\alias{exportPython}
\title{The function that permit to exchange the data from R to Python}
\usage{
exportPython(.Object)
}
\arguments{
\item{.Object}{is the reference of the object in the method}

\item{split}{What to split on.}
}
\value{
A dataframe of the components of the model.
}
\description{
The function that permit to exchange the data from R to Python
}
