% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{getDF}
\alias{getDF}
\title{Get the dataframe saved in the model}
\usage{
getDF(.Object)
}
\arguments{
\item{.Object}{is the reference of the object in the method}
}
\value{
The dataframe
}
\description{
Get the dataframe saved in the model
}
