% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{getModel}
\alias{getModel}
\title{Return the model from the wrapper}
\usage{
getModel(name)
}
\arguments{
\item{.Object}{is the reference of the object in the method}
}
\value{
The str model.
}
\description{
Return the model from the wrapper
}
