% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{initialize}
\alias{initialize}
\title{Inizialize the wrapper object for stR}
\usage{
initialize(.Object, ...)
}
\arguments{
\item{.Object}{is the reference of the object in the method}
}
\value{
The updated object.
}
\description{
Inizialize the wrapper object for stR
}
