% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{leftJ}
\alias{leftJ}
\title{Make the left join}
\usage{
leftJ(.Object, pathFile, nameColumn)
}
\arguments{
\item{.Object}{is the reference of the object in the method}

\item{pathFile}{specific the path that contains a dataframe to be joined}

\item{nameColumn}{specific the name of the equal name of column to do the left join}
}
\value{
The new stR wrapper with the dataframe updated
}
\description{
Make the left join
}
