% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{loadDF}
\alias{loadDF}
\title{Load the dataframe from a specific location}
\usage{
loadDF(.Object, pathDF)
}
\arguments{
\item{.Object}{is the reference of the object in the method}

\item{pathDF}{contains the path to the dataframe}
}
\value{
The new str object with the dataframe updated
}
\description{
Load the dataframe from a specific location
}
