% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{saveModel}
\alias{saveModel}
\title{Save the model in the RDS format}
\usage{
saveModel(model.str, name)
}
\arguments{
\item{name}{the name of the model saved}

\item{.Object}{is the reference of the object in the method}
}
\value{
Nothing
}
\description{
Save the model in the RDS format
}
