% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sunspot}
\alias{sunspot}
\title{Sunspot}
\format{A tibble with 270 rows and 11 variables:
\describe{
  \item{spot_0}{transformed sunspot}
  \item{spot_1}{transformed sunspot of lag one}
  \item{spot_2}{transformed sunspot of lag one}
  \item{spot_3}{transformed sunspot of lag one}
  \item{spot_4}{transformed sunspot of lag one}
  \item{spot_5}{transformed sunspot of lag one}
  \item{spot_6}{transformed sunspot of lag one}
  \item{spot_7}{transformed sunspot of lag one}
  \item{spot_8}{transformed sunspot of lag one}
  \item{spot_9}{transformed sunspot of lag one}
  \item{spot_10}{transformed sunspot of lag one}
}}
\usage{
sunspot
}
\description{
A dataset containing a balanced panel data of annual observations
over the period 1973-1987 (15 years) for 560 US firms for the variables described below.
}
\details{
The structure of the dataset is such that the time index runs “fast”, while the firm index runs “slow”;
that is, first all 14 observations for the first firm are given, then the 14 observations for the second firm, etc.

Since we used one year lagged variables of "vala", "debta", "cfa" and "cfa" as regressors,
the records in 1973 are skipped.

All values are nominal and millions of dollars except where otherwise noted. Stocks are end of year.
}
\keyword{datasets}
