% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{train}
\alias{train}
\title{Train the model}
\usage{
train(.Object, gapCV, nFold, value, trace)
}
\arguments{
\item{.Object}{is the reference of the object in the method}
}
\value{
The trained model.
}
\description{
Train the model
}
