% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStR.r
\name{updateDF}
\alias{updateDF}
\title{Set the new dataframe in the object}
\usage{
updateDF(.Object, newDF)
}
\arguments{
\item{.Object}{is the reference of the object in the method}

\item{newDF}{the path to a new dataframe.}
}
\value{
The stR object with the new dataframe.
}
\description{
Set the new dataframe in the object
}
