% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{version}
\alias{version}
\title{Show the version number of some information.}
\usage{
version()
}
\description{
This function shows the version number and some information of the package.
}
\author{
Yukai Yang, \email{yukai.yang@statistik.uu.se}
}
\keyword{utils}
