% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sunspot}
\alias{sunspot}
\title{Transformed Wolf annual sunspot numbers for the years 1710-1979}
\format{A tibble with 270 rows and 11 variables:
\describe{
  \item{spot_0}{transformed sunspot}
  \item{spot_1}{transformed sunspot, lag one}
  \item{spot_2}{transformed sunspot, lag two}
  \item{spot_3}{transformed sunspot, lag three}
  \item{spot_4}{transformed sunspot, lag four}
  \item{spot_5}{transformed sunspot, lag five}
  \item{spot_6}{transformed sunspot, lag six}
  \item{spot_7}{transformed sunspot, lag seven}
  \item{spot_8}{transformed sunspot, lag eight}
  \item{spot_9}{transformed sunspot, lag nine}
  \item{spot_10}{transformed sunspot, lag ten}
}}
\source{
\url{http://sidc.oma.be/html/sunspot.html}
}
\usage{
sunspot
}
\description{
A dataset containing the transformed Wolf annual sunspot numbers for the years 1710-1979.
}
\details{
Each column of the data matrix is a lagged transformed sunspot observations from lag order 0 to 10.

The data were transformed by using the formula
\deqn{y_t = 2 \left\{ (1 + x_t)^{1/2} -1 \right\}}
see Ghaddar and Tong (1981)
}
\section{References}{

Ghaddar, D. K. and Tong, H. (1981) Data transformation and self-exciting threshold autoregression, Applied Statistics, 30, 238–48.
}

\keyword{datasets}
